+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 ### Changed
 - Open an already opened object return the same reference
+- Change key type of waitingPull from DeltaCRDT to CObjectUId
 ### Deprecated
 ### Removed
 ### Fixed